# services/api/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastUI Sales API",
    "description": "Production-grade API for lead prospecting, sales pipeline, and exports.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/prospects": {
      "get": {
        "tags": [
          "prospects"
        ],
        "summary": "List Prospects",
        "description": "Fetches businesses that have not yet been promoted to active sales leads.",
        "operationId": "list_prospects_v1_prospects_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Max records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max records to return"
          },
          {
            "name": "qualification_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by qualification status",
              "title": "Qualification Status"
            },
            "description": "Filter by qualification status"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search across business name, category, and city",
              "title": "Search"
            },
            "description": "Search across business name, category, and city"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Field to sort by",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort direction",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessResponse"
                  },
                  "title": "Response List Prospects V1 Prospects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "prospects"
        ],
        "summary": "Update Prospects",
        "description": "Bulk updates qualification status for selected prospects.",
        "operationId": "update_prospects_v1_prospects_patch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkQualifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkQualifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/exports": {
      "post": {
        "tags": [
          "Exports"
        ],
        "summary": "Enqueue CSV Export",
        "description": "Enqueues an asynchronous CSV export for prospects, leads, or businesses.",
        "operationId": "create_export_v1_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ExportCreateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exports/{export_id}": {
      "get": {
        "tags": [
          "Exports"
        ],
        "summary": "Get Export Status",
        "description": "Checks progress status and download URL of an export job.",
        "operationId": "get_export_status_v1_exports__export_id__get",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exports/{export_id}/download": {
      "get": {
        "tags": [
          "Exports"
        ],
        "summary": "Download Exported CSV File",
        "description": "Streams the generated CSV export with formula injection defense and Excel UTF-8 BOM.",
        "operationId": "download_export_v1_exports__export_id__download_get",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/notifications/subscribe": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Subscribe to Push Notifications",
        "description": "Registers or updates a browser Web Push subscription.",
        "operationId": "subscribe_push_notifications_v1_notifications_subscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/unsubscribe": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Unsubscribe from Push Notifications",
        "description": "Removes a push subscription endpoint.",
        "operationId": "unsubscribe_push_notifications_v1_notifications_unsubscribe_delete",
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Endpoint"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/test": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Send Test Push Notification",
        "description": "Sends a test push notification to all devices registered by current user.",
        "operationId": "send_test_notification_v1_notifications_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestNotificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Send Test Notification V1 Notifications Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/broadcast": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Broadcast Push Notification",
        "description": "Broadcasts a push notification to specific users or all registered accounts.",
        "operationId": "broadcast_notification_v1_notifications_broadcast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastNotificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastNotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkQualifyRequest": {
        "properties": {
          "business_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Business Ids"
          },
          "qualification_status": {
            "type": "string",
            "title": "Qualification Status"
          }
        },
        "type": "object",
        "required": [
          "business_ids",
          "qualification_status"
        ],
        "title": "BulkQualifyRequest"
      },
      "BulkQualifyResponse": {
        "properties": {
          "updated_count": {
            "type": "integer",
            "title": "Updated Count"
          },
          "business_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Business Ids"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "updated_count",
          "business_ids"
        ],
        "title": "BulkQualifyResponse"
      },
...
      "BusinessResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "has_whatsapp": {
            "type": "boolean",
            "title": "Has Whatsapp",
            "default": false
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "website_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {}
            ],
            "title": "Website Status",
            "default": "unknown"
          },
          "qualification_status": {
            "type": "string",
            "title": "Qualification Status",
            "default": "unqualified"
          },
          "is_lead": {
            "type": "boolean",
            "title": "Is Lead",
            "default": false
          },
          "lead_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Id"
          },
          "pipeline_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Stage"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "medium"
          },
          "signal": {
            "type": "string",
            "title": "Signal",
            "default": "warm"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "default": 50
          },
          "google_maps_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Maps Url"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "google_place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Place Id"
          },
          "last_outreach_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outreach At"
          },
          "last_contacted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Contacted At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "business_name"
        ],
        "title": "BusinessResponse",
        "description": "Canonical representation of a business entity with lead/pipeline enrichment."
      },
...
      "ExportCreateRequest": {
        "properties": {
          "export_type": {
            "$ref": "#/components/schemas/ExportType",
            "description": "Target entity to export",
            "default": "prospects"
          },
          "scope": {
            "$ref": "#/components/schemas/ExportScope",
            "description": "Scope of export: selected, filtered, or all",
            "default": "all"
          },
          "record_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Ids",
            "description": "List of IDs to export when scope is 'selected'"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExportFilterParams"
              },
              {
                "type": "null"
              }
            ],
            "description": "Active filters to preserve in export"
          }
        },
        "type": "object",
        "title": "ExportCreateRequest"
      },
      "ExportCreateResponse": {
        "properties": {
          "export_id": {
            "type": "string",
            "title": "Export Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "export_id",
          "message"
        ],
        "title": "ExportCreateResponse"
      },
...
      "ExportScope": {
        "type": "string",
        "enum": [
          "selected",
          "filtered",
          "all"
        ],
        "title": "ExportScope"
      },
      "ExportStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "export_type": {
            "type": "string",
            "title": "Export Type",
            "default": "prospects"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent"
          },
          "records_processed": {
            "type": "integer",
            "title": "Records Processed"
          },
          "total_records": {
            "type": "integer",
            "title": "Total Records"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "progress_percent",
          "records_processed",
          "total_records"
        ],
        "title": "ExportStatusResponse"
      },
      "ExportType": {
        "type": "string",
        "enum": [
          "prospects",
          "leads",
          "businesses"
        ],
        "title": "ExportType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PushSubscriptionCreate": {
        "properties": {
          "endpoint": {
            "type": "string",
            "title": "Endpoint",
            "description": "Unique push service subscription endpoint URL"
          },
          "keys": {
            "$ref": "#/components/schemas/PushSubscriptionKeys"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent",
            "description": "Optional browser user-agent string"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "PushSubscriptionCreate"
      },
...
      "PushSubscriptionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "endpoint",
          "created_at"
        ],
        "title": "PushSubscriptionResponse"
      },
...
    }
  }
}